Working to fix issues with integrating into aprs-is-demo.
