+++ package.json
@@ -40,7 +40,6 @@
     "@istanbuljs/nyc-config-typescript": "^0.1.3",
     "@types/chai": "^4.2.4",
     "@types/mocha": "^5.2.7",
-    "@types/node": "^12.12.3",
     "chai": "^4.2.0",
     "coveralls": "^3.0.7",
     "mocha": "^6.2.2",
@@ -48,5 +47,8 @@
     "source-map-support": "^0.5.16",
     "ts-node": "^8.4.1",
     "typescript": "^3.6.4"
+  },
+  "dependencies": {
+    "@types/node": "^12.12.5"
   }
 }
